% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vocab}
\alias{Vocab}
\title{Vocabulary and Education}
\format{This data frame contains the following columns:
\describe{
\item{year}{Year of the survey.}
\item{sex}{Sex of the respondent, Female or Male.}
\item{education}{Education, in years.}
\item{vocabulary}{Vocabulary test score: number correct on a 10-word test.}
}}
\source{
HNational Opinion Research Center General Social Survey. GSS Cumulative Datafile 1972-2016, originally downloaded from \code{\link{http://gss.norc.org/}}.
}
\usage{
Vocab
}
\description{
The Vocab data frame has 30,351 rows and 4 columns. The observations are respondents to U.S. General Social Surveys, 1972-2016.
}
\references{
Fox, J. (2016) Applied Regression Analysis and Generalized Linear Models, Third Edition. Sage.

Fox, J. and Weisberg, S. (2019) An R Companion to Applied Regression, Third Edition, Sage.
}
\keyword{datasets}
