% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writereports.R
\name{write.report}
\alias{write.report}
\title{Write Formatted Output to a *.docx File}
\usage{
write.report(
  object,
  path,
  tablefont = "Arial",
  tablefontsize = 10,
  digits = NULL
)
}
\arguments{
\item{object}{a R object from an aca analysis.}

\item{path}{the location to save the docx file.}

\item{tablefont}{Tdesignates the font to use in report creation. A list of available font families can be obtained with the  \code{\link{sys_fonts()}} command.}

\item{tablefontsize}{designates the font size to use in report creation.}

\item{digits}{he number of decimal digits to be reported in each table. If this value is not specified, but rounding was performed on the aca object, the number of rounded digits will be used. If no value is specified, 4 digits will be used by default.}
}
\value{

}
\description{
The write.report function will produce an MS Word document using the \code{\link{officer}}.
}
